chore: update gulp config to fix icons scripts
