+++ packages/icons/package.json
@@ -20,7 +20,7 @@
   "module": "vaadin-iconset.js",
   "type": "module",
   "scripts": {
-    "icons": "gulp icons"
+    "icons": "gulp icons --gulpfile gulpfile.cjs"
   },
   "files": [
     "iconset.js",
