+++ packages/rich-text-editor/package.json
@@ -21,7 +21,7 @@
   "module": "vaadin-rich-text-editor.js",
   "type": "module",
   "scripts": {
-    "icons": "gulp icons"
+    "icons": "gulp icons --gulpfile gulpfile.cjs"
   },
   "files": [
     "src",
