+++ packages/vaadin-material-styles/gulpfile.cjs
@@ -2,11 +2,10 @@
 'use strict';
 
 const gulp = require('gulp');
-const sort = require('gulp-sort');
 const iconfont = require('gulp-iconfont');
 const fs = require('fs');
-const svgpath = require('svgpath');
 const svgmin = require('gulp-svgmin');
+const sort = require('gulp-sort');
 
 /**
  * Normalize file sort order across platforms (OS X vs Linux, maybe others).
@@ -24,42 +23,7 @@ function sortIconFilesNormalized(file1, file2) {
   return file1.replace(/-/g, '~').localeCompare(file2.replace(/-/g, '~'), 'en-US');
 }
 
-function createCopyright() {
-  return `/**
- * @license
- * Copyright (c) 2017 - 2022 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */`;
-}
-
-function createIconset(folder, filenames, idPrefix = '') {
-  let output = `<svg xmlns="http://www.w3.org/2000/svg">\n<defs>\n`;
-  filenames.forEach((filename) => {
-    // Skip non-svg files
-    if (filename.indexOf('.svg') === -1) {
-      return;
-    }
-
-    const content = fs.readFileSync(folder + filename, 'utf-8');
-    const path = content.match(/<path( fill-rule="evenodd" clip-rule="evenodd")* d="([^"]*)"/);
-    if (path) {
-      const newPath = new svgpath(path[2])
-        .scale(1000 / 24, 1000 / 24)
-        .round(0)
-        .toString();
-      const name = filename.replace('.svg', '').replace(/\s/g, '-').toLowerCase();
-      const attrs = path[1] !== undefined ? path[1] : '';
-      output += `<g id="${idPrefix}${name}"><path d="${newPath}"${attrs}></path></g>\n`;
-    } else {
-      throw new Error(`Unexpected SVG content: ${filename}`);
-    }
-  });
-
-  output += `</defs>\n</svg>`;
-  return output;
-}
-
-gulp.task('icons', () => {
+gulp.task('icons', (done) => {
   const folder = 'icons/svg/';
   let glyphs;
 
@@ -90,48 +54,6 @@ gulp.task('icons', () => {
     )
     .pipe(gulp.dest(folder))
     .on('finish', () => {
-      // Iron-iconset-svg
-      fs.readdir(folder, (err, filenames) => {
-        if (err) {
-          console.error(err);
-          return;
-        }
-
-        filenames.sort(sortIconFilesNormalized);
-
-        const ironIcons = `${createCopyright()}
-import '@polymer/iron-iconset-svg/iron-iconset-svg.js';
-import './version.js';
-
-const template = document.createElement('template');
-
-template.innerHTML = \`<iron-iconset-svg size="1000" name="lumo">
-${createIconset(folder, filenames)}
-</iron-iconset-svg>\`;\n\ndocument.head.appendChild(template.content);\n`;
-
-        fs.writeFile('iconset.js', ironIcons, (err) => {
-          if (err) {
-            return console.error(err);
-          }
-        });
-
-        const vaadinIcons = `${createCopyright()}
-import '@vaadin/icon/vaadin-iconset.js';
-import './version.js';
-
-const template = document.createElement('template');
-
-template.innerHTML = \`<vaadin-iconset name="lumo" size="1000">
-${createIconset(folder, filenames, 'lumo:')}
-</vaadin-iconset>\`;\n\ndocument.head.appendChild(template.content);\n`;
-
-        fs.writeFile('vaadin-iconset.js', vaadinIcons, (err) => {
-          if (err) {
-            return console.error(err);
-          }
-        });
-      });
-
       // Icon font
       gulp
         .src(`${folder}*.svg`)
@@ -144,12 +66,10 @@ ${createIconset(folder, filenames, 'lumo:')}
         )
         .pipe(
           iconfont({
-            fontName: 'lumo-icons',
+            fontName: 'material-icons',
             formats: ['woff'],
-            fontHeight: 1000,
-            ascent: 850,
-            descent: 150,
-            fixedWidth: true,
+            fontHeight: 2400,
+            descent: 400,
             normalize: true,
             timestamp: 1, // Truthy!
           }),
@@ -161,11 +81,13 @@ ${createIconset(folder, filenames, 'lumo:')}
         .pipe(gulp.dest('.'))
         .on('finish', () => {
           // Generate base64 version of the font
-          const lumoIconsWoff = fs.readFileSync('lumo-icons.woff');
-
+          const materialIconsWoff = fs.readFileSync('material-icons.woff');
           // Write the output to font-icons.js
-          let output = createCopyright();
-          output += `
+          let output = `/**
+ * @license
+ * Copyright (c) 2017 - 2022 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
 import './version.js';
 
 const template = document.createElement('template');
@@ -173,8 +95,8 @@ const template = document.createElement('template');
 template.innerHTML = \`
   <style>
     @font-face {
-      font-family: 'lumo-icons';
-      src: url(data:application/font-woff;charset=utf-8;base64,${lumoIconsWoff.toString('base64')}) format('woff');
+      font-family: 'material-icons';
+      src: url(data:application/font-woff;charset=utf-8;base64,${materialIconsWoff.toString('base64')}) format('woff');
       font-weight: normal;
       font-style: normal;
     }
@@ -184,7 +106,7 @@ template.innerHTML = \`
           glyphs.forEach((g) => {
             const name = g.name.replace(/\s/g, '-').toLowerCase();
             const unicode = `\\\\${g.unicode[0].charCodeAt(0).toString(16)}`;
-            output += `      --lumo-icons-${name}: "${unicode}";\n`;
+            output += `      --material-icons-${name}: "${unicode}";\n`;
           });
           output += `    }
   </style>
@@ -206,10 +128,11 @@ document.head.appendChild(template.content);
           });
 
           // Cleanup
-          fs.unlink('lumo-icons.woff', (err) => {
+          fs.unlink('material-icons.woff', (err) => {
             if (err) {
               return console.error(err);
             }
+            done();
           });
         });
     });
